+++ module-service/src/main/java/com/checkping/service/question/link/QuestionLinkService.java
@@ -1,10 +1,10 @@
 package com.checkping.service.question.link;
 
-import com.checkping.dto.question.link.TaskBoardLinkRequest;
-import com.checkping.dto.question.link.TaskBoardLinkResponse.TaskBoardLinkDto;
+import com.checkping.dto.question.link.QuestionLinkRequest;
+import com.checkping.dto.question.link.QuestionLinkResponse.TaskBoardLinkDto;
 
 public interface QuestionLinkService {
 
     TaskBoardLinkDto register(Long taskBoardId,
-        TaskBoardLinkRequest.RegisterDto request);
+        QuestionLinkRequest.RegisterDto request);
 }
